scalable pagination done, lets built a full memory toolchain (include an abstractor)

+++ inc/drivers/vga.h
@@ -8,8 +8,7 @@
 #define VGA_HEIGHT 25
 
 extern uint64 __KERNEL_VIRT_BASE;
-#define VGA_BUFFER_ADDRESS ((uint16 *)(0xB8000 + &__KERNEL_VIRT_BASE))
-// #define VGA_BUFFER_ADDRESS ((uint16 *)(0xB8000))
+#define VGA_BUFFER_ADDRESS ((uint16 *)((uint64)0xB8000 + (uint64)(&__KERNEL_VIRT_BASE)))
 
 /* struct use to managing vga */
 struct vga
